Moved cache purgers out of ApplicationContext/ directory
