+++ src/Infrastructure/Cache/AllOfTheCachePurger.php
@@ -2,7 +2,7 @@
 
 declare( strict_types = 1 );
 
-namespace WMDE\Fundraising\Frontend\ApplicationContext\Infrastructure;
+namespace WMDE\Fundraising\Frontend\Infrastructure\Cache;
 
 use Doctrine\Common\Cache\Cache;
 use Doctrine\Common\Cache\FlushableCache;
